Title: 계단 오르기, Time: 124ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 250829/계단 오르기/climbing-stairs.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int[] dp = new int[1001];
+        long[] dp = new long[1001];
         dp[2] = 1;
         dp[3] = 1;
         for (int i = 4; i < 1001; i++) {
